+++ amethyst_renderer/src/hidden.rs
@@ -1,17 +1,14 @@
 use amethyst_core::specs::{
     prelude::Component,
-    storage::{
-        FlaggedStorage,
-        NullStorage,
-    },
+    storage::{FlaggedStorage, NullStorage},
 };
 
 /// Hidden mesh component
-/// Useful for Ui-Elements that should stay loaded, but not visible.
+/// Useful for entities, that should not be rendered, but stay loaded in memory.
 #[derive(Clone, Debug, Default)]
 pub struct Hidden;
 
-impl Component for Hidden{
+impl Component for Hidden {
     //type Storage = NullStorage<Self>;
     type Storage = FlaggedStorage<Self, NullStorage<Self>>;
 }